Refactored Layout.tsx
